Handle parsed xml content as inputs
